fix-irq-flags-size

+++ drivers/rtc/rtc-ds1511.c
@@ -185,7 +185,7 @@ ds1511_wdog_disable(void)
 ds1511_rtc_set_time(struct device *dev, struct rtc_time *rtc_tm)
 {
 	u8 mon, day, dow, hrs, min, sec, yrs, cen;
-	unsigned int flags;
+	unsigned long flags;
 
 	/*
 	 * won't have to change this for a while
@@ -249,7 +249,7 @@ ds1511_rtc_set_time(struct device *dev, struct rtc_time *rtc_tm)
 ds1511_rtc_read_time(struct device *dev, struct rtc_time *rtc_tm)
 {
 	unsigned int century;
-	unsigned int flags;
+	unsigned long flags;
 
 	spin_lock_irqsave(&ds1511_lock, flags);
 	rtc_disable_update();
